requirements.txt
absl-py==0.12.0
amqp==5.0.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asciinema==2.0.2
asciitree==0.3.3
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
autobahn==20.12.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.3.0
bleach==3.3.0
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
celery==5.0.5
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
chime==0.6.0
chrome-gnome-shell==0.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cligj==0.7.1
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==3.3.1
cupshelpers==1.0
cycler==0.10.0
dblatex===0.3.11py3
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.4
django-rest==0.8.6
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
enum34==1.1.10
fasteners==0.14.1
filelock==3.0.12
Fiona==1.8.18
flatbuffers==1.12
future==0.18.2
gast==0.3.3
geopandas==0.8.1
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.6.4
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-tuner==1.0.2
keyring==18.0.1
kiteconnect==3.9.0
kiwisolver==1.3.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libiio==0.19
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
munch==2.5.0
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
nltk==3.6.2
notebook==6.3.0
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.8
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
Pygments==2.8.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyproj==3.0.0.post1
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
schedule==1.1.0
scikit-learn==0.24.1
scipy==1.6.2
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.1
simplejson==3.16.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
terminaltables==3.1.0
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.2.1
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zope.interface==5.2.0
